Bluetooth: Mesh: Add iterations param script
This commit adds an `iterations` parameter to the generic script
that runs network simulations. This allows users to specify the number of
iterations for each test run, providing flexibility in testing
different network conditions and behaviors.

Removed scripts for network2 and network3 tests, as they are not needed.
Only one sample for network1 is kept.

Added violin plots for three scenarios as well.

Signed-off-by: Omkar Kulkarni <omkar.kulkarni@nordicsemi.no>